backend/requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.2.5
django-browser-reload==1.11.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
Markdown==3.4.4
mypy==1.5.1
mypy-extensions==1.0.0
pydocstyle==6.3.0
python-decouple==3.8
pytz==2023.3.post1
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
